Extract app settings helpers
