+++ src/App.vue
@@ -59,6 +59,7 @@ import type { ComposerDraftPayload, ThreadComposerExposed } from './components/c
 import type { GitCommitOption, LocalDirectoryEntry, TelegramStatus, ThreadTerminalQuickCommand, WorktreeBranchOption } from './api/codexGateway'
 import { getFreeModeStatus, setFreeMode, setFreeModeCustomKey, setCustomProvider } from './api/codexGateway'
 import { buildExportFileName, buildThreadMarkdown } from './appExportMarkdown'
+import { applyDarkMode, buildDictationLanguageOptions, buildTerminalHeaderQuickCommands, CHAT_WIDTH_PRESETS, loadBoolPref, loadChatWidthPref, loadDarkModePref, loadDictationLanguagePref, loadInProgressSendModePref, loadTerminalStoredQuickCommands, normalizeTerminalQuickCommandValue, normalizeToWhisperLanguage, saveTerminalStoredQuickCommands, type ChatWidthMode, type DirectoryTryItemPayload, type TerminalHeaderQuickCommand, type ThreadTerminalPanelExposed } from './appSettingsSupport'
 import { getPathLeafName, getPathParent, isProjectlessChatPath, normalizePathForUi } from './pathUtils.js'
 
 const ThreadConversation = defineAsyncComponent(() => import('./components/content/ThreadConversation.vue'))
@@ -83,7 +84,6 @@ function onProviderSelectChange(event: Event): void {
 
 const SIDEBAR_COLLAPSED_STORAGE_KEY = 'codex-web-local.sidebar-collapsed.v1'
 const ACCOUNTS_SECTION_COLLAPSED_STORAGE_KEY = 'codex-web-local.accounts-section-collapsed.v1'
-const TERMINAL_QUICK_COMMAND_STORAGE_KEY = 'codex-web-local.terminal-quick-commands.v1'
 const TOGGLE_TERMINAL_COMMAND_VALUE = '__toggle_terminal__'
 const worktreeName = import.meta.env.VITE_WORKTREE_NAME ?? 'unknown'
 const appVersion = import.meta.env.VITE_APP_VERSION ?? 'unknown'
@@ -97,156 +97,6 @@ const SETTINGS_HELP = {
   dictationLanguage: t('Choose transcription language or keep auto-detect.'),
 } as const
 
-type ChatWidthMode = 'standard' | 'wide' | 'extra-wide'
-
-type TerminalHeaderQuickCommand = {
-  label: string
-  value: string
-  custom?: boolean
-  usageCount: number
-  lastUsedAt: number
-  sourceIndex?: number
-}
-
-type ThreadTerminalPanelExposed = {
-  runQuickCommand: (command: string, custom?: boolean) => Promise<void>
-}
-
-type DirectoryTryItemPayload = {
-  kind: 'app' | 'plugin' | 'skill' | 'composio'
-  name: string
-  displayName: string
-  skillPath?: string
-  prompt?: string
-  attachedSkills?: Array<{ name: string; path: string }>
-}
-
-type ChatWidthPreset = {
-  label: string
-  columnMax: string
-  cardMax: string
-}
-
-const CHAT_WIDTH_PRESETS: Record<ChatWidthMode, ChatWidthPreset> = {
-  standard: {
-    label: 'Standard',
-    columnMax: '45rem',
-    cardMax: '76ch',
-  },
-  wide: {
-    label: 'Wide',
-    columnMax: '72rem',
-    cardMax: '88ch',
-  },
-  'extra-wide': {
-    label: 'Extra wide',
-    columnMax: '96rem',
-    cardMax: '96ch',
-  },
-}
-
-const WHISPER_LANGUAGES: Record<string, string> = {
-  en: 'english',
-  zh: 'chinese',
-  de: 'german',
-  es: 'spanish',
-  ru: 'russian',
-  ko: 'korean',
-  fr: 'french',
-  ja: 'japanese',
-  pt: 'portuguese',
-  tr: 'turkish',
-  pl: 'polish',
-  ca: 'catalan',
-  nl: 'dutch',
-  ar: 'arabic',
-  sv: 'swedish',
-  it: 'italian',
-  id: 'indonesian',
-  hi: 'hindi',
-  fi: 'finnish',
-  vi: 'vietnamese',
-  he: 'hebrew',
-  uk: 'ukrainian',
-  el: 'greek',
-  ms: 'malay',
-  cs: 'czech',
-  ro: 'romanian',
-  da: 'danish',
-  hu: 'hungarian',
-  ta: 'tamil',
-  no: 'norwegian',
-  th: 'thai',
-  ur: 'urdu',
-  hr: 'croatian',
-  bg: 'bulgarian',
-  lt: 'lithuanian',
-  la: 'latin',
-  mi: 'maori',
-  ml: 'malayalam',
-  cy: 'welsh',
-  sk: 'slovak',
-  te: 'telugu',
-  fa: 'persian',
-  lv: 'latvian',
-  bn: 'bengali',
-  sr: 'serbian',
-  az: 'azerbaijani',
-  sl: 'slovenian',
-  kn: 'kannada',
-  et: 'estonian',
-  mk: 'macedonian',
-  br: 'breton',
-  eu: 'basque',
-  is: 'icelandic',
-  hy: 'armenian',
-  ne: 'nepali',
-  mn: 'mongolian',
-  bs: 'bosnian',
-  kk: 'kazakh',
-  sq: 'albanian',
-  sw: 'swahili',
-  gl: 'galician',
-  mr: 'marathi',
-  pa: 'punjabi',
-  si: 'sinhala',
-  km: 'khmer',
-  sn: 'shona',
-  yo: 'yoruba',
-  so: 'somali',
-  af: 'afrikaans',
-  oc: 'occitan',
-  ka: 'georgian',
-  be: 'belarusian',
-  tg: 'tajik',
-  sd: 'sindhi',
-  gu: 'gujarati',
-  am: 'amharic',
-  yi: 'yiddish',
-  lo: 'lao',
-  uz: 'uzbek',
-  fo: 'faroese',
-  ht: 'haitian creole',
-  ps: 'pashto',
-  tk: 'turkmen',
-  nn: 'nynorsk',
-  mt: 'maltese',
-  sa: 'sanskrit',
-  lb: 'luxembourgish',
-  my: 'myanmar',
-  bo: 'tibetan',
-  tl: 'tagalog',
-  mg: 'malagasy',
-  as: 'assamese',
-  tt: 'tatar',
-  haw: 'hawaiian',
-  ln: 'lingala',
-  ha: 'hausa',
-  ba: 'bashkir',
-  jw: 'javanese',
-  su: 'sundanese',
-  yue: 'cantonese',
-}
 
 const {
   projectGroups,
@@ -422,7 +272,7 @@ const chatWidth = ref<ChatWidthMode>(loadChatWidthPref())
 const dictationClickToToggle = ref(loadBoolPref(DICTATION_CLICK_TO_TOGGLE_KEY, false))
 const dictationAutoSend = ref(loadBoolPref(DICTATION_AUTO_SEND_KEY, true))
 const dictationLanguage = ref(loadDictationLanguagePref())
-const dictationLanguageOptions = computed(() => buildDictationLanguageOptions())
+const dictationLanguageOptions = computed(() => buildDictationLanguageOptions(t, dictationLanguage.value, typeof navigator !== 'undefined' ? (navigator.languages ?? []) : []))
 const showFirstLaunchPluginsCard = ref(false)
 const freeModeEnabled = ref(false)
 const freeModeLoading = ref(false)
@@ -827,6 +677,7 @@ const existingFolderFilteredEntries = computed(() => {
   )
 })
 const darkModeMediaQuery = typeof window !== 'undefined' ? window.matchMedia('(prefers-color-scheme: dark)') : null
+const handleDarkModeMediaChange = (): void => applyDarkMode(darkMode.value)
 const chatWidthLabel = computed(() => t(CHAT_WIDTH_PRESETS[chatWidth.value].label))
 const terminalShortcutLabel = computed(() => {
   if (typeof navigator !== 'undefined' && /mac|iphone|ipad|ipod/i.test(navigator.platform)) {
@@ -837,22 +688,7 @@ const terminalShortcutLabel = computed(() => {
 const terminalCommandPlaceholder = computed(() => (
   isComposerTerminalOpen.value ? t('Terminal') : t('Open terminal')
 ))
-const terminalHeaderQuickCommands = computed<TerminalHeaderQuickCommand[]>(() => {
-  const storedByValue = new Map(terminalStoredQuickCommands.value.map((command) => [command.value, command]))
-  const combined: TerminalHeaderQuickCommand[] = [
-    ...terminalProjectQuickCommands.value.map((command, index) => ({
-      label: command.label,
-      value: command.value,
-      usageCount: 0,
-      lastUsedAt: 0,
-      ...(storedByValue.get(command.value) ?? {}),
-      custom: false,
-      sourceIndex: index,
-    })),
-  ]
-  return combined
-    .sort(compareTerminalQuickCommands)
-})
+const terminalHeaderQuickCommands = computed<TerminalHeaderQuickCommand[]>(() => buildTerminalHeaderQuickCommands(terminalProjectQuickCommands.value, terminalStoredQuickCommands.value))
 const terminalHeaderDropdownOptions = computed(() => [
   { label: isComposerTerminalOpen.value ? t('Hide terminal') : t('Open terminal'), value: TOGGLE_TERMINAL_COMMAND_VALUE },
   ...terminalHeaderQuickCommands.value.map((command) => ({ label: command.label, value: command.value })),
@@ -892,8 +728,8 @@ onMounted(() => {
   window.visualViewport?.addEventListener('resize', updateVisualViewportState)
   window.visualViewport?.addEventListener('scroll', updateVisualViewportState)
   updateVisualViewportState()
-  applyDarkMode()
-  darkModeMediaQuery?.addEventListener('change', applyDarkMode)
+  applyDarkMode(darkMode.value)
+  darkModeMediaQuery?.addEventListener('change', handleDarkModeMediaChange)
   void initialize()
   void loadHomeDirectory()
   void loadFirstLaunchPluginsCardPreference()
@@ -915,7 +751,7 @@ onUnmounted(() => {
   window.removeEventListener('resize', updateVisualViewportState)
   window.visualViewport?.removeEventListener('resize', updateVisualViewportState)
   window.visualViewport?.removeEventListener('scroll', updateVisualViewportState)
-  darkModeMediaQuery?.removeEventListener('change', applyDarkMode)
+  darkModeMediaQuery?.removeEventListener('change', handleDarkModeMediaChange)
   if (accountStatePollTimer !== null) {
     window.clearInterval(accountStatePollTimer)
     accountStatePollTimer = null
@@ -1748,74 +1584,6 @@ function recordHeaderTerminalCommandUse(command: string): void {
   saveTerminalStoredQuickCommands(next)
 }
 
-function normalizeTerminalQuickCommandValue(value: string): string {
-  return value.trim().replace(/\s+/g, ' ')
-}
-
-function compareTerminalQuickCommands(first: TerminalHeaderQuickCommand, second: TerminalHeaderQuickCommand): number {
-  if (second.usageCount !== first.usageCount) return second.usageCount - first.usageCount
-  if (second.lastUsedAt !== first.lastUsedAt) return second.lastUsedAt - first.lastUsedAt
-  const firstSource = typeof first.sourceIndex === 'number' ? first.sourceIndex : Number.MAX_SAFE_INTEGER
-  const secondSource = typeof second.sourceIndex === 'number' ? second.sourceIndex : Number.MAX_SAFE_INTEGER
-  return firstSource - secondSource
-}
-
-function loadTerminalStoredQuickCommands(): TerminalHeaderQuickCommand[] {
-  if (typeof window === 'undefined') return []
-  try {
-    const raw = window.localStorage.getItem(TERMINAL_QUICK_COMMAND_STORAGE_KEY)
-    if (!raw) return []
-    const parsed = JSON.parse(raw) as unknown
-    if (!Array.isArray(parsed)) return []
-    const seen = new Set<string>()
-    const commands: TerminalHeaderQuickCommand[] = []
-    for (const row of parsed) {
-      const record = row !== null && typeof row === 'object' && !Array.isArray(row)
-        ? row as Record<string, unknown>
-        : null
-      const value = normalizeTerminalQuickCommandValue(readTerminalString(record?.value))
-      if (!value || seen.has(value)) continue
-      seen.add(value)
-      commands.push({
-        label: readTerminalString(record?.label) || value,
-        value,
-        custom: record?.custom !== false,
-        usageCount: readTerminalPositiveInteger(record?.usageCount),
-        lastUsedAt: readTerminalPositiveInteger(record?.lastUsedAt),
-      })
-    }
-    return commands
-  } catch {
-    return []
-  }
-}
-
-function saveTerminalStoredQuickCommands(commands: TerminalHeaderQuickCommand[]): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(
-    TERMINAL_QUICK_COMMAND_STORAGE_KEY,
-    JSON.stringify(commands.map((command) => ({
-      label: command.label,
-      value: command.value,
-      custom: command.custom === true,
-      usageCount: command.usageCount,
-      lastUsedAt: command.lastUsedAt,
-    }))),
-  )
-}
-
-function readTerminalString(value: unknown): string {
-  return typeof value === 'string' ? value : ''
-}
-
-function readTerminalPositiveInteger(value: unknown): number {
-  if (typeof value === 'number' && Number.isFinite(value)) return Math.max(0, Math.trunc(value))
-  if (typeof value === 'string') {
-    const parsed = Number(value)
-    if (Number.isFinite(parsed)) return Math.max(0, Math.trunc(parsed))
-  }
-  return 0
-}
 
 function onTerminalFocusChange(focused: boolean): void {
   isTerminalInputFocused.value = focused
@@ -1995,7 +1763,6 @@ function onEditQueuedMessage(messageId: string): void {
   composer.hydrateDraft(payload)
   removeQueuedMessage(messageId)
 }
-
 
 function scheduleMobileConversationJumpToLatest(): void {
   if (!isMobile.value || isHomeRoute.value) return
@@ -2613,7 +2380,6 @@ function onImplementPlan(payload: { turnId: string }): void {
   void sendMessageToSelectedThread('Implement', [], [], 'steer', [], undefined, 'default')
 }
 
-
 function onExportChat(): void {
   if (isHomeRoute.value || isSkillsRoute.value || isAutomationsRoute.value || typeof document === 'undefined') return
   if (!selectedThread.value || filteredMessages.value.length === 0) return
@@ -2630,32 +2396,6 @@ function onExportChat(): void {
   window.setTimeout(() => URL.revokeObjectURL(objectUrl), 0)
 }
 
-function loadBoolPref(key: string, fallback: boolean): boolean {
-  if (typeof window === 'undefined') return fallback
-  const v = window.localStorage.getItem(key)
-  if (v === null) return fallback
-  return v === '1'
-}
-
-function loadDarkModePref(): 'system' | 'light' | 'dark' {
-  if (typeof window === 'undefined') return 'system'
-  const v = window.localStorage.getItem(DARK_MODE_KEY)
-  if (v === 'light' || v === 'dark') return v
-  return 'system'
-}
-
-function loadInProgressSendModePref(): 'steer' | 'queue' {
-  if (typeof window === 'undefined') return 'steer'
-  const v = window.localStorage.getItem(IN_PROGRESS_SEND_MODE_KEY)
-  if (v === 'steer' || v === 'queue') return v
-  return 'queue'
-}
-
-function loadChatWidthPref(): ChatWidthMode {
-  if (typeof window === 'undefined') return 'standard'
-  const value = window.localStorage.getItem(CHAT_WIDTH_KEY)
-  return value === 'standard' || value === 'wide' || value === 'extra-wide' ? value : 'standard'
-}
 
 function toggleSendWithEnter(): void {
   sendWithEnter.value = !sendWithEnter.value
@@ -2672,7 +2412,7 @@ function cycleDarkMode(): void {
   const idx = order.indexOf(darkMode.value)
   darkMode.value = order[(idx + 1) % order.length]
   window.localStorage.setItem(DARK_MODE_KEY, darkMode.value)
-  applyDarkMode()
+  applyDarkMode(darkMode.value)
 }
 
 function cycleChatWidth(): void {
@@ -2691,7 +2431,6 @@ function toggleDictationAutoSend(): void {
   dictationAutoSend.value = !dictationAutoSend.value
   window.localStorage.setItem(DICTATION_AUTO_SEND_KEY, dictationAutoSend.value ? '1' : '0')
 }
-
 
 async function onProviderChange(provider: string): Promise<void> {
   if (freeModeLoading.value) return
@@ -2861,72 +2600,7 @@ function onDictationLanguageChange(nextValue: string): void {
   window.localStorage.setItem(DICTATION_LANGUAGE_KEY, value)
 }
 
-function loadDictationLanguagePref(): string {
-  if (typeof window === 'undefined') return 'auto'
-  const value = window.localStorage.getItem(DICTATION_LANGUAGE_KEY)?.trim() || 'auto'
-  const normalized = normalizeToWhisperLanguage(value)
-  return normalized || 'auto'
-}
 
-function buildDictationLanguageOptions(): Array<{ value: string; label: string }> {
-  const options: Array<{ value: string; label: string }> = [{ value: 'auto', label: t('Auto-detect') }]
-  const seen = new Set<string>(['auto'])
-  function formatLanguageLabel(value: string): string {
-    const languageName = WHISPER_LANGUAGES[value] || value
-    const title = languageName.charAt(0).toUpperCase() + languageName.slice(1)
-    return `${title} (${value})`
-  }
-
-  for (const raw of typeof navigator !== 'undefined' ? (navigator.languages ?? []) : []) {
-    const value = normalizeToWhisperLanguage(raw)
-    if (!value || seen.has(value)) continue
-    seen.add(value)
-    options.push({
-      value,
-      label: `Preferred: ${formatLanguageLabel(value)}`,
-    })
-  }
-
-  for (const value of Object.keys(WHISPER_LANGUAGES)) {
-    if (seen.has(value)) continue
-    seen.add(value)
-    options.push({
-      value,
-      label: formatLanguageLabel(value),
-    })
-  }
-
-  const current = dictationLanguage.value.trim()
-  if (current && !seen.has(current)) {
-    options.push({
-      value: current,
-      label: formatLanguageLabel(current),
-    })
-  }
-
-  return options
-}
-
-function normalizeToWhisperLanguage(raw: string): string {
-  const value = raw.trim().toLowerCase()
-  if (!value || value === 'auto') return ''
-  if (value in WHISPER_LANGUAGES) return value
-  const base = value.split('-')[0] ?? value
-  if (base in WHISPER_LANGUAGES) return base
-  return ''
-}
-
-function applyDarkMode(): void {
-  const root = document.documentElement
-  if (darkMode.value === 'dark') {
-    root.classList.add('dark')
-  } else if (darkMode.value === 'light') {
-    root.classList.remove('dark')
-  } else {
-    const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches
-    root.classList.toggle('dark', prefersDark)
-  }
-}
 
 function loadSidebarCollapsed(): boolean {
   if (typeof window === 'undefined') return false
@@ -3201,7 +2875,6 @@ watch(
   },
   { immediate: true },
 )
-
 
 watch(isMobile, (mobile) => {
   if (mobile && !isSidebarCollapsed.value) {
